+++ app/vmui/packages/vmui/src/components/Configurators/QueryEditor/LogsQL/useFetchLogsQLOptions.tsx
@@ -10,6 +10,7 @@ import { AUTOCOMPLETE_LIMITS } from "../../../../constants/queryAutocomplete";
 import { LogsFiledValues } from "../../../../api/types";
 import { useLogsDispatch, useLogsState } from "../../../../state/logsPanel/LogsStateContext";
 import { useTenant } from "../../../../hooks/useTenant";
+import { generateQuery } from "./utils";
 
 type FetchDataArgs = {
   urlSuffix: string;
@@ -113,7 +114,7 @@ export const useFetchLogsQLOptions = (contextData?: ContextData) => {
       urlSuffix: "field_names",
       setter: setFieldNames,
       type: ContextType.FilterName,
-      params: getQueryParams({ query: "*" })
+      params: getQueryParams({ query: contextData?.queryBeforeIncompleteFilter || "*" })
     });
 
     return () => abortControllerRef.current?.abort();
@@ -132,7 +133,7 @@ export const useFetchLogsQLOptions = (contextData?: ContextData) => {
       urlSuffix: "field_values",
       setter: setFieldValues,
       type: ContextType.FilterValue,
-      params: getQueryParams({ query: "*", field: contextData.filterName })
+      params: getQueryParams({ query: generateQuery(contextData), field: contextData.filterName })
     });
 
     return () => abortControllerRef.current?.abort();